cdc: remove leftovers of old CDC management generation functions
